+++ Shared/Position.h
@@ -7,17 +7,17 @@ namespace ecs
 	class Position: public AComponent
 	{
 	public:
-		Position(int x, int y, int z);
+		Position(float x, float y, float z);
 		Position(const Position& pos);
 		Position& operator=(const Position& pos);
 		~Position() = default;
 
-		void set(int x, int y, int z);
+		void set(float x, float y, float z);
 		void set(const Position& pos);
 
-		void add(int x, int y, int z);
+		void add(float x, float y, float z);
 		void add(const Position& pos);
-		void sub(int x, int y, int z);
+		void sub(float x, float y, float z);
 		void sub(const Position& pos);
 
 		Position& operator+=(const Position& pos);
@@ -26,15 +26,15 @@ namespace ecs
 		Position operator+(const Position& pos) const;
 		Position operator-(const Position& pos) const;
 
-		int getX() const;
-		int getY() const;
-		int getZ() const;
+		float getX() const;
+		float getY() const;
+		float getZ() const;
 
 		void dump() const;
 
 	private:
-		int m_iX;
-		int m_iY;
-		int m_iZ;
+		float m_iX;
+		float m_iY;
+		float m_iZ;
 	};
 }
